+++ tests/testthat/test_resample.R
@@ -94,7 +94,7 @@ test_that("empty train/predict sets", {
 })
 
 test_that("conditions are returned", {
-  expect_true(all(c("warnings", "errors") %in% names(rr$score(conditions = TRUE))))
+  expect_true(all(c("warnings", "errors") %chin% names(rr$score(conditions = TRUE))))
 })
 
 test_that("save/load roundtrip", {
@@ -212,13 +212,13 @@ test_that("marshaling works when store_models is FALSE", {
     resample(task, learner, resampling, store_models = FALSE, unmarshal = TRUE)
   })
   expect_resample_result(rr)
-  expect_true(is.null(rr$learners[[1]]$model))
+  expect_null(rr$learners[[1]]$model)
 
   rr1 = with_future(future::sequential, {
     resample(task, learner, resampling, store_models = FALSE, unmarshal = TRUE)
   })
   expect_resample_result(rr1)
-  expect_true(is.null(rr1$learners[[1]]$model))
+  expect_null(rr1$learners[[1]]$model)
 })